refactor(provider): use adc instead of admin-api

+++ test/e2e/scaffold/scaffold.go
@@ -464,9 +464,12 @@ func (s *Scaffold) afterEach() {
 			}
 		}
 	}
-	// if the test case is successful, just delete namespace
-	err := k8s.DeleteNamespaceE(s.t, s.kubectlOptions, s.namespace)
-	Expect(err).NotTo(HaveOccurred(), "deleting namespace "+s.namespace)
+
+	if os.Getenv("TSET_KEEP") != "true" {
+		// if the test case is successful, just delete namespace
+		err := k8s.DeleteNamespaceE(s.t, s.kubectlOptions, s.namespace)
+		Expect(err).NotTo(HaveOccurred(), "deleting namespace "+s.namespace)
+	}
 
 	for _, f := range s.finalizers {
 		runWithRecover(f)
